review-feedback: BaseEntity에서 id제외하기
