+++ src/main/java/qna/domain/Answer.java
@@ -1,6 +1,9 @@
 package qna.domain;
 
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.Lob;
 import javax.persistence.Table;
 import qna.NotFoundException;
@@ -9,11 +12,13 @@ import qna.UnAuthorizedException;
 import java.util.Objects;
 @Entity
 @Table(name = "answer" )
-public class Answer extends BaseEntity{
+public class Answer extends BaseAuditingEntity {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
     private Long writerId;
     private Long questionId;
-
     @Lob
     private String contents;
     private boolean deleted = false;
@@ -27,7 +32,7 @@ public class Answer extends BaseEntity{
     }
 
     public Answer(Long id, User writer, Question question, String contents) {
-        setId(id);
+        this.id = id;
         if (Objects.isNull(writer)) {
             throw new UnAuthorizedException();
         }
@@ -39,6 +44,14 @@ public class Answer extends BaseEntity{
         this.writerId = writer.getId();
         this.questionId = question.getId();
         this.contents = contents;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
     }
 
     public boolean isOwner(User writer) {
@@ -84,7 +97,7 @@ public class Answer extends BaseEntity{
     @Override
     public String toString() {
         return "Answer{" +
-                "id=" + getId() +
+                "id=" + id +
                 ", writerId=" + writerId +
                 ", questionId=" + questionId +
                 ", contents='" + contents + '\'' +
