+++ src/main/java/qna/domain/Question.java
@@ -2,12 +2,18 @@ package qna.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.Lob;
 import javax.persistence.Table;
 
 @Entity
 @Table(name="question")
-public class Question extends BaseEntity{
+public class Question extends BaseAuditingEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
     @Column(length = 100, nullable = false)
     private String title;
     @Lob
@@ -24,7 +30,7 @@ public class Question extends BaseEntity{
     }
 
     public Question(Long id, String title, String contents) {
-        setId(id);
+        this.id=id;
         this.title = title;
         this.contents = contents;
     }
@@ -32,6 +38,14 @@ public class Question extends BaseEntity{
     public Question writeBy(User writer) {
         this.writerId = writer.getId();
         return this;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
     }
 
     public boolean isOwner(User writer) {
@@ -77,7 +91,7 @@ public class Question extends BaseEntity{
     @Override
     public String toString() {
         return "Question{" +
-                "id=" + getId() +
+                "id=" + id +
                 ", title='" + title + '\'' +
                 ", contents='" + contents + '\'' +
                 ", writerId=" + writerId +
